level_5/requirements.txt
fastapi
uvicorn
numpy
google-genai
websockets
python-dotenv
google-adk
a2a-sdk
aiokafka
